Update README.md
The 'Meditation' app is here to give you a break from a busy day. It helps you feel less stressed and tired by taking you to a calm and quiet place with sounds of nature. It's a helpful way to take care of your mind and feel better."

Build using vanilla JavaScript, HTML, CSS.